requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
openpyxl
mysql-connector-python


